airflow/requirements.txt
# Core dependencies for Airflow DAGs
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
mlflow==2.16.0
requests>=2.32.0
python-dotenv==1.0.0
joblib==1.3.2
pyyaml==6.0.1
statsmodels==0.14.1
ydata-profiling==4.10.0
evidently==0.4.40
